refactor: molecules 오타 수정
