+++ src/components/molecules/Input/Input.stories.tsx
@@ -2,13 +2,17 @@ import { action } from '@storybook/addon-actions';
 import type { Meta, StoryObj } from '@storybook/react';
 import { useEffect, useState } from 'react';
 
-import Textarea from './Textarea';
+import Input from './Input';
 
 const meta = {
-  title: 'components/atoms/Textarea',
-  component: Textarea,
+  title: 'components/molecules/Input',
+  component: Input,
   tags: ['autodocs'],
   argTypes: {
+    inputSize: {
+      control: 'radio',
+      options: ['sm', 'md', 'lg'],
+    },
     label: {
       control: 'text',
     },
@@ -35,15 +39,17 @@ const meta = {
       options: ['error', 'warning', 'info', 'success'],
     },
     onChange: { action: 'changed' },
+    onClear: { action: 'cleared' },
   },
-} satisfies Meta<typeof Textarea>;
+} satisfies Meta<typeof Input>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
-    label: '소개글',
+    inputSize: 'md',
+    label: '이메일',
     placeholder: 'Placeholder',
     value: '',
     required: true,
@@ -53,22 +59,28 @@ export const Default: Story = {
     errorStatus: 'error',
   },
   render: (args) => {
-    const [text, setText] = useState(args.value);
+    const [inputValue, setInputValue] = useState(args.value);
 
     useEffect(() => {
-      setText(args.value);
+      setInputValue(args.value);
     }, [args.value]);
 
-    const handleChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-      setText(e.target.value);
+    const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+      setInputValue(e.target.value);
       action('changed')(e.target.value);
     };
 
+    const handleClear = () => {
+      setInputValue('');
+      action('cleared')();
+    };
+
     return (
-      <Textarea
+      <Input
         {...args}
-        value={text}
+        value={inputValue}
         onChange={handleChange}
+        onClear={handleClear}
       />
     );
   },
