+++ src/components/molecules/Textarea/Textarea.stories.tsx
@@ -2,17 +2,13 @@ import { action } from '@storybook/addon-actions';
 import type { Meta, StoryObj } from '@storybook/react';
 import { useEffect, useState } from 'react';
 
-import Input from './Input';
+import Textarea from './Textarea';
 
 const meta = {
-  title: 'components/atoms/Input',
-  component: Input,
+  title: 'components/molecules/Textarea',
+  component: Textarea,
   tags: ['autodocs'],
   argTypes: {
-    inputSize: {
-      control: 'radio',
-      options: ['sm', 'md', 'lg'],
-    },
     label: {
       control: 'text',
     },
@@ -39,17 +35,15 @@ const meta = {
       options: ['error', 'warning', 'info', 'success'],
     },
     onChange: { action: 'changed' },
-    onClear: { action: 'cleared' },
   },
-} satisfies Meta<typeof Input>;
+} satisfies Meta<typeof Textarea>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
-    inputSize: 'md',
-    label: '이메일',
+    label: '소개글',
     placeholder: 'Placeholder',
     value: '',
     required: true,
@@ -59,28 +53,22 @@ export const Default: Story = {
     errorStatus: 'error',
   },
   render: (args) => {
-    const [inputValue, setInputValue] = useState(args.value);
+    const [text, setText] = useState(args.value);
 
     useEffect(() => {
-      setInputValue(args.value);
+      setText(args.value);
     }, [args.value]);
 
-    const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-      setInputValue(e.target.value);
+    const handleChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+      setText(e.target.value);
       action('changed')(e.target.value);
     };
 
-    const handleClear = () => {
-      setInputValue('');
-      action('cleared')();
-    };
-
     return (
-      <Input
+      <Textarea
         {...args}
-        value={inputValue}
+        value={text}
         onChange={handleChange}
-        onClear={handleClear}
       />
     );
   },
